+++ hook/lua/ScenarioPlatoonAI.lua
@@ -739,3 +739,13 @@ EngineerBuildAI = function(self)
             aiBrain:DisbandPlatoon(self)
         end
     end
+	
+--- Enables Stealth on platoon's units
+---@param platoon Platoon
+function PlatoonEnableStealth(platoon)
+    for _, unit in platoon:GetPlatoonUnits() do
+        if not unit.Dead and unit:TestToggleCaps('RULEUTC_StealthToggle') then
+            unit:SetScriptBit('RULEUTC_StealthToggle', false)
+        end
+    end
+end